first method is ready

+++ Methods_of_calculus/Laba5_ApproxCalcOfIntegralsQFHighestDegree/Laba5_ApproxCalcOfIntegralsQFHighestDegree/UserInterface.cs
@@ -85,12 +85,12 @@ namespace Laba5_ApproxCalcOfIntegralsQFHighestDegree
         {
             Console.Write("\nВведите левую границу отрезка: ");
             Console.ForegroundColor = ConsoleColor.Yellow;
-            leftBorder = int.Parse(Console.ReadLine());
+            leftBorder = double.Parse(Console.ReadLine());
             Console.ResetColor();
 
             Console.Write("\nВведите правую границу отрезка: ");
             Console.ForegroundColor = ConsoleColor.Yellow;
-            rightBorder = int.Parse(Console.ReadLine());
+            rightBorder = double.Parse(Console.ReadLine());
             while (rightBorder <= leftBorder)
             {
                 Console.ForegroundColor = ConsoleColor.Red;
